fix: PR review comment 반영
